requirements.txt
requests~=2.28.1
pytz~=2022.1
icalendar~=4.1.0
Pillow~=9.2.0
recurring-ical-events
schedule
